add act.js
